Make /showcase/<slug> a real page
It was a <meta refresh> stub bouncing to /showcase?project=<slug>, so a
shared link cost a redirect, the content wasn't indexable, and analytics
couldn't see it — a pageview fired there would race the navigation.

Extract the modal's layout into ProjectDetail, shared by the modal and a
standalone page wrapped in StarlightPage for the site chrome. The grid now
pushes /showcase/<slug>, so the URL is the same whether you clicked a card
or arrived cold, and Plausible records it as an ordinary pageview with no
custom event or goal to configure.

Keep ?project= links working (the old stub advertised them as canonical):
they open the project and replaceState to the real path. Sync the modal to
popstate so back/forward no longer leaves it stranded open.
